requirements.txt
# Core Data Processing
pandas==2.1.4
numpy==1.26.2
pyspark==3.5.0

# Machine Learning
scikit-learn==1.3.2
pyod==1.1.3
# hdbscan==0.8.33  # Optional - requires C++ build tools, uncomment if needed for Phase 3

# Deep Learning (for Phase 3 - Autoencoders) - Optional
# tensorflow==2.15.0  # Uncomment for Phase 3
# keras==2.15.0  # Uncomment for Phase 3
# torch==2.1.2  # Alternative to TensorFlow

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
folium==0.15.1

# Interpretability
shap==0.44.0

# Geospatial
geopy==2.4.1

# Utilities
tqdm==4.66.1
pyyaml==6.0.1
python-dateutil==2.8.2
pytz==2023.3

# Google Drive Integration
gdown==4.7.1

# Jupyter (using notebook instead of full jupyter to avoid long path issues)
notebook==7.0.6
ipykernel==6.27.1
ipywidgets==8.1.1
jupyter-client==8.6.0

# Experiment Tracking (Optional)
# mlflow==2.9.2

# Data Version Control (Optional)
# dvc==3.36.1
